+++ CHANGELOG.md
@@ -5,7 +5,25 @@ All notable changes to Anki Taskbar will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+
+## [unreleased]
+
+### Added
+- Random Session Arrangement feature
+- Session suffle Feature
+
+### Fixed
+- 
+
+### Changed
+- 
+
+### Removed
+- 
+
+
+
+## [v1.2.0] -2026-01-01
 
 ### Added
 - Enhanced mouse event tracking with detailed coordinate logging